extended AVISimpleIndex.aIndex
git-svn-id: svn://localhost/video_utilities/trunk@407 682e2775-9c8c-9945-981b-715035ffa45a

+++ dv_utils/dv_dump/avi.cpp
@@ -310,7 +310,7 @@ int AVIFile::GetTotalFrames() const
  
 */
 
-void AVIFile::ParseList(int parent)
+void AVIFile::ParseList(size_t parent)
 {
     FOURCC      type;
     FOURCC      name;
